refactor: remove useless build code

+++ scripts/postbuild.js
@@ -12,7 +12,7 @@ if (fs.pathExistsSync(outputDir)) {
 
 list.forEach(({ modPath, mod }) => {
   if (!mod.includes('theme') && !mod.includes('admin')) return;
-  const buildDirName = fs.existsSync(`${modPath}/dist`) ? 'dist' : 'build';
-  const buildPath = path.join(modPath, `./${buildDirName}`);
+  const buildPath = path.join(modPath, './dist');
+  if (!fs.existsSync(buildPath)) return;
   fs.copySync(buildPath, `${outputDir}/${path.basename(modPath)}`);
 });
